// Web/src/app/models/PestVM.ts
export class PestVM{
    ID_pest: number;
    name: string;
    description: string;
    type: string;
    severity: string;
    spread: string;
    symptoms: string;
    prevention: string;
    remedy: string;
    company: number;
}

export class SearchPestVM{
    TotalRows: number;
    ID_pest: number;
    name: string;
    description: string;
    type: string;
    severity: string;
    spread: string;
    symptoms: string;
    prevention: string;
    remedy: string;
    company: number;
}